projects/sinan/2_digitador_sinan_api/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
alembic
sqlalchemy
requests